+++ karma.conf.js
@@ -120,6 +120,10 @@ module.exports = function(config) {
               }
             }
           }
+        ],
+        noParse: [
+          // avoid munging internal benchmark script magic
+          /benchmark/
         ]
       },
       node: {