+++ tests/test-common.js
@@ -3,7 +3,6 @@
  */
 /* eslint-disable indent */
 const EarlReport = require('./earl-report');
-const benchmark = require('benchmark');
 const join = require('join-path-js');
 const rdfCanonize = require('rdf-canonize');
 const {prependBase} = require('../lib/url');
@@ -13,6 +12,7 @@ module.exports = function(options) {
 'use strict';
 
 const assert = options.assert;
+const benchmark = options.benchmark;
 const jsonld = options.jsonld;
 
 const manifest = options.manifest || {
@@ -573,8 +573,8 @@ function addTest(manifest, test, tests) {
           throw Error('Unknown test type: ' + test.type);
         }
 
-        let benchResult = null
-        if(options.benchmark) {
+        let benchResult = null;
+        if(options.benchmarkOptions) {
           const result = await runBenchmark({
             test,
             fn,