+++ tests/test.js
@@ -25,6 +25,7 @@
  * Copyright (c) 2011-2019 Digital Bazaar, Inc. All rights reserved.
  */
 const assert = require('chai').assert;
+const benchmark = require('benchmark');
 const common = require('./test-common');
 const fs = require('fs-extra');
 const jsonld = require('..');
@@ -84,13 +85,13 @@ if(process.env.JSONLD_TESTS) {
   entries.push(path.resolve(_top, 'tests/node-document-loader-tests.js'));
 }
 
-let benchmark = null;
+let benchmarkOptions = null;
 if(process.env.JSONLD_BENCHMARK) {
-  benchmark = {};
+  benchmarkOptions = {};
   if(!(['1', 'true'].includes(process.env.JSONLD_BENCHMARK))) {
     process.env.JSONLD_BENCHMARK.split(',').forEach(pair => {
       const kv = pair.split('=');
-      benchmark[kv[0]] = kv[1];
+      benchmarkOptions[kv[0]] = kv[1];
     });
   }
 }
@@ -100,6 +101,7 @@ const options = {
     path
   },
   assert,
+  benchmark,
   jsonld,
   exit: code => process.exit(code),
   earl: {
@@ -110,7 +112,7 @@ const options = {
   verboseSkip: process.env.VERBOSE_SKIP === 'true',
   bailOnError: process.env.BAIL === 'true',
   entries,
-  benchmark,
+  benchmarkOptions,
   readFile: filename => {
     return fs.readFile(filename, 'utf8');
   },